# Opitmize LayerNormOp