… caffe2/operators/layer_norm_op.h