Summary: Opitmize LayerNormOp

Differential Revision: D12398163