@BIT-silence 2018-10-26
Some benchmark.
LayerNormGradientOp on CPU, Tensor Shape = [100, 100, 2000], axis = 1, 841.64ms -> 57.48ms.
LayerNormOp on GPU, Tensor Shape = [100, 1000, 2000], axis = 1, 1439.45ms -> 16.18ms
LauyerNormGradientOp on GPU, Tensor Shape = [100, 1000, 2000], axis = 1, 2148.43ms -> 22.27ms